+++ dev/openchpl_data-model.sql
@@ -2231,7 +2231,7 @@ CREATE TABLE openchpl.surveillance_requirement (
 		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE,
 	CONSTRAINT certification_criterion_fk FOREIGN KEY (certification_criterion_id) 
 		REFERENCES openchpl.certification_criterion (certification_criterion_id) 
-		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE,		
+		MATCH FULL ON DELETE CASCADE ON UPDATE CASCADE,		
 	CONSTRAINT result_fk FOREIGN KEY (result_id) 
 		REFERENCES openchpl.surveillance_result (id) 
 		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE		
@@ -2265,10 +2265,66 @@ CREATE TABLE openchpl.surveillance_nonconformity (
 		MATCH FULL ON DELETE CASCADE ON UPDATE CASCADE,
 	CONSTRAINT certification_criterion_fk FOREIGN KEY (certification_criterion_id) 
 		REFERENCES openchpl.certification_criterion (certification_criterion_id) 
-		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE,
+		MATCH FULL ON DELETE CASCADE ON UPDATE CASCADE,
 	CONSTRAINT nonconformity_status_fk FOREIGN KEY (nonconformity_status_id) 
 		REFERENCES openchpl.nonconformity_status (id) 
 		MATCH FULL ON DELETE SET NULL ON UPDATE CASCADE	
+);
+
+CREATE TABLE openchpl.pending_surveillance (
+	id bigserial not null,
+	certified_product_unique_id varchar(30),
+	start_date date,
+	end_date date,
+	type_value varchar(30),
+	randomized_sites_used integer, 
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT pending_surveillance_pk PRIMARY KEY (id)	
+);
+
+CREATE TABLE openchpl.pending_surveillance_requirement (
+	id bigserial not null,
+	pending_surveillance_id bigint not null,
+	type_value varchar(50),
+	requirement varchar(1024),
+	result_value varchar(30),
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT pending_surveillance_requirement_id PRIMARY KEY (id),
+	CONSTRAINT pending_surveillance_fk FOREIGN KEY (pending_surveillance_id) 
+		REFERENCES openchpl.pending_surveillance (id) 
+		MATCH FULL ON DELETE CASCADE ON UPDATE CASCADE	
+);
+
+CREATE TABLE openchpl.pending_surveillance_nonconformity (
+	id bigserial not null,
+	pending_surveillance_requirement_id bigint not null,
+	nonconformtiy_type varchar(1024), 
+	nonconformity_status varchar(15),
+	date_of_determination date,
+	corrective_action_plan_approval_date date,
+	corrective_action_start_date date,
+	corrective_action_must_complete_date date,
+	corrective_action_end_date date,
+	summary varchar(2048), 
+	findings text, 
+	sites_passed integer, 
+	total_sites integer,
+	developer_explanation text,
+	resolution text,
+	creation_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_date timestamp NOT NULL DEFAULT NOW(),
+	last_modified_user bigint NOT NULL,
+	deleted bool NOT NULL DEFAULT false,
+	CONSTRAINT pending_surveillance_nonconformity_pk PRIMARY KEY (id),
+	CONSTRAINT pending_surveillance_requirement_fk FOREIGN KEY (pending_surveillance_requirement_id) 
+		REFERENCES openchpl.pending_surveillance_requirement (id) 
+		MATCH FULL ON DELETE CASCADE ON UPDATE CASCADE
 );
 
 -- Table: openchpl.ehr_certification_id