+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@ _Date TBD_
 * Added script to remove three specific duplicate certified products
 * Added product owner history relationship
 * Add surveillance tables to the database 
+* Add pending surveillance tables to the database.
 
 ---